Adicionando as classes service e restcontroller das entidades Cliente e Funcionario

+++ java/seboDoJoao/src/main/java/br/com/interdisciplinar/sebodojoao/repository/FuncionarioRepository.java
@@ -3,6 +3,5 @@ package br.com.interdisciplinar.sebodojoao.repository;
 import br.com.interdisciplinar.sebodojoao.model.Funcionario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FuncionarioRepository extends JpaRepository<Funcionario, Integer> {
-
+public interface FuncionarioRepository extends JpaRepository<Funcionario, Long> {
 }
